Reflect can-be-strong split in DHO structure
